requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
pytz==2020.5
sqlparse==0.4.1
factory-boy==3.2.0
PyJWT~=1.7.1